refactor: couponAvailability 유틸 함수로 쿠폰 유효성 검증 로직 분리
